OZ-68: Added bootstrap class for each mode
